backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.0
django-cors-headers==4.3.1
torch==2.9.0
torchvision==0.24.0
scikit-learn==1.3.2
Pillow==10.1.0
numpy==1.26.3
nibabel==5.1.0
scipy==1.11.4
scikit-image==0.22.0
